Add script to generate a dependency graph.
It relies on data available on the uploader's machine, but could be
adjusted to double check what's happening in the Sources file once all
packages are accepted into the archive.

This graph makes it possible to trigger source-only uploads in the right
order once NEW has been cleared.

The script (ab)uses the fact the relevant packages are following the
same naming convention (binary := source + '-dev'). It could easily be
made smarter, but that's probably not required at this stage.